Complete report refactor
